backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
mysql-connector-python==8.2.0
langchain-community==0.3.27
langchain==0.3.26
langchain-openai==0.3.27
openai==1.3.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
aiohttp==3.12.14
python-multipart==0.0.6